Move VirtIOSocket methods to new VsockManager trait for VsockConnectionManager
This commit moves VirtIOSocket methods to a new VsockManager trait for internal
use and redefines VsockConnectionManager to use these trait methods. This will
allow most of VsockConncetionManager and VirtIOSocket methods to share a common
implementation (using trait method defaults) regardless of whether the
VsockConnectionManager is used for the driver or device side. For vsock this
makes sense because devices and drivers are symmetric in that they poll/send
buffers on opposite queues (i.e. driver polls rx vqueue and sends to tx vqueue,
device polls tx vqueue and sends to rx vqueue). While the VsockConnectionManager
changes are not breaking in practice, the VirtIOSocket change currently is
breaking (downstream users would need to import VsockManager trait) to minimize
the diff. Both sets of changes could be made non-breaking by 1. redefining
VsockConnectionManager as a struct with a single field instead of a type alias
and 2. keeping VirtIOSocket methods which call the default implementations in
the VsockManager trait.