Make VirtQueueLayout generic over DmaMemory and add device-side constructors
This commit redefines VirtQueueLayout in terms of a type implementing DmaMemory
intead of being specific to the driver-side Dma type. Methods which are
applicable to both sides are defined for VirtQueueLayout<D> while the existing
constructors are kept specific to VirtQueueLayout<Dma<H>>. It also adds new
constructors for VirtQueueLayout<DeviceDma<H>>. VirtQueueLayout is internal to
the queue module so this should not be a breaking change.