Add VirtIOSocketManager trait and impl for VirtIOSocket
This trait will only be used internally by VsockConnectionManager so it can
handle the driver and device sides with a common implementation.
VirtIOSocketManager requires methods for the local CID, polling the receive
virtqueue and sending packets on the transmit virtqueue. It then provides most
of the existing VirtIOSocket methods using default implementations. Since this
trait is not intended for use outside the crate this commit also redefines the
VirtIOSocket methods in terms of the ones provided by the trait, using fully
qualified syntax to disambiguate methods.